Started work on project organization
Decided to divide the project into many smaller crates, like for example for the architecture dependant code. Also updated the 'qemu' bash script to copy the executable to a dedicated disk image folder, with the added benefit of it automatically booting the executable, instead of needing one to manually execute it from the EFI shell. Updated the 'README' as well